chore(windows): move kmtip_guids.h and testkeymanfunctioning.h
