+++ windows/include/kmtip_guids.h
