+++ windows/src/engine/keyman32/Keyman32.vcxproj.filters
@@ -246,7 +246,7 @@
     <ClInclude Include="kmprocessactions.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\global\inc\kmtip_guids.h">
+    <ClInclude Include="..\..\..\include\kmtip_guids.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
