+++ windows/src/engine/keyman32/keymanengine.h
@@ -264,7 +264,7 @@ void keybd_shift(LPINPUT pInputs, int* n, BOOL isReset, LPBYTE const kbd);
 
 #include "testkeymanfunctioning.h"
 #include "keynames.h"
-#include "kmtip_guids.h"
+#include "..\..\..\include\kmtip_guids.h"
 
 #include "crc32.h"
