+++ windows/src/engine/keyman64/keyman64.vcxproj
@@ -277,7 +277,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\keyman32\keymanengine.h" />
-    <ClInclude Include="..\..\global\inc\kmtip_guids.h" />
+    <ClInclude Include="..\..\..\include\kmtip_guids.h" />
     <ClInclude Include="..\keyman32\addins.h" />
     <ClInclude Include="..\keyman32\appint\aiTIP.h" />
     <ClInclude Include="..\keyman32\appint\aiWin2000Unicode.h" />
@@ -314,7 +314,7 @@
     <ClInclude Include="..\keyman32\serialkeyeventserver.h" />
     <ClInclude Include="..\keyman32\SharedBuffers.h" />
     <ClInclude Include="..\keyman32\syskbd.h" />
-    <ClInclude Include="..\..\global\inc\testkeymanfunctioning.h" />
+    <ClInclude Include="..\keyman32\testkeymanfunctioning.h" />
     <ClInclude Include="..\..\global\inc\unicode.h" />
     <ClInclude Include="..\keyman32\vkscancodes.h" />
   </ItemGroup>
